Adds a check for empty project identifiers
I was excited to see that traffic to the widget had picked up
significantly over the last couple of days (~30%) and then I noticed
that the server was under heavy load... unfortunately it turns out not
to be a significant increase in usage, but due to the new BBCode on
Minecraft Forum some users have accidentally embedded empty widgets in
their posts ([mod][/mod]) and every single request has been triggered a
request to curse.com which locked up the server. At least it was an easy
fix.